# Conflicts:
#	.gitignore
#	users/static/css/styles.css
#	users/templates/users/activity.html
#	users/templates/users/chat.html
#	users/templates/users/home.html

+++ users/views.py
@@ -137,7 +137,17 @@ def password_reset_request(request):  # pragma: no cover
         form = ResetPasswordRequestForm(request.POST)
         if form.is_valid():
             username = form.cleaned_data.get("username")
-            user = User.objects.get(username=username)
+            try:
+                user = User.objects.get(username=username)
+            except Exception:
+                form = ResetPasswordRequestForm()
+                form.errors["username"] = "User with that NetID does not exist."
+                return render(
+                    request,
+                    "users/passwordreset/password_reset_request_form.html",
+                    {"form": form},
+                )
+
             to_email = username + "@nyu.edu"
             current_site = get_current_site(request)
             mail_subject = "Reset your NYUnite Account Password!"
@@ -175,16 +185,9 @@ def password_reset(request, uidb64, token):  # pragma: no cover
         if request.method == "POST":
             form = ResetPasswordForm(request.POST)
             if form.is_valid():
-                error_bool, password = form.clean_password()
-                if error_bool:
-                    return render(
-                        request,
-                        "users/passwordreset/password_reset_form.html",
-                        {"form": form},
-                    )
-                else:
-                    user.set_password(password)
-                    user.save()
+                password = form.cleaned_data.get("new_password1")
+                user.set_password(password)
+                user.save()
 
                 return render(
                     request, "users/passwordreset/password_reset_successful_signin.html"
@@ -206,7 +209,7 @@ def profile_setup(request):
 
     try:
         prof = Profile.objects.get(user=request.user)
-    except Exception:
+    except Exception:  # pragma: no cover
         prof = Profile(user=request.user)
         prof.save()
 
@@ -215,7 +218,7 @@ def profile_setup(request):
         if form.is_valid():
             ans = form.save()
 
-            if "image" in request.FILES:
+            if "image" in request.FILES:  # pragma: no cover
                 ans.image = request.FILES["image"]
 
             ans.save()
@@ -231,7 +234,7 @@ def update_profile(request):
 
     try:
         prof = Profile.objects.get(user=request.user)
-    except Exception:
+    except Exception:  # pragma: no cover
         prof = Profile(user=request.user)
         prof.save()
 
@@ -240,7 +243,7 @@ def update_profile(request):
         if form.is_valid():
             ans = form.save()
 
-            if "image" in request.FILES:
+            if "image" in request.FILES:  # pragma: no cover
                 ans.image = request.FILES["image"]
 
             ans.save()
@@ -422,10 +425,10 @@ def accept_request(user, id):
     user2 = from_user
     user1.profile.friends.add(user2.profile)
     user2.profile.friends.add(user1.profile)
-    if FriendRequest.objects.filter(from_user=user, to_user=from_user).first():
-        request_rev = FriendRequest.objects.filter(
-            from_user=user, to_user=from_user
-        ).first()
+    request_rev = FriendRequest.objects.filter(
+        from_user=user, to_user=from_user
+    ).first()
+    if request_rev:
         request_rev.delete()
     frequest.delete()
 
@@ -441,7 +444,8 @@ def accept_request_query(request):
 def decline_request(user, id):
     from_user = User.objects.get(id=id)
     frequest = FriendRequest.objects.filter(from_user=from_user, to_user=user).first()
-    frequest.delete()
+    if frequest:
+        frequest.delete()
 
 
 @login_required
@@ -457,7 +461,7 @@ def search(request):
     search_query, query_set = get_search(request)
     friend_list = request.user.profile.friends.all()
     friend_list_ids = []
-    for friend in friend_list:
+    for friend in friend_list:  # pragma: no cover
         friend_list_ids.append(friend.id)
 
     return render(
@@ -502,7 +506,6 @@ def approve_suggestion(request):
         accepter = User.objects.get(id=request.user.id)
         acceptee = User.objects.get(id=user_id)
         accepter.profile.seen_users.add(acceptee.profile)
-        print(accepter.profile.seen_users)
 
         # Send a friend request
         send_friend_request(request.user, user_id)
@@ -712,218 +715,3 @@ def activity_search(request):
 def activity(request):
     # event_data = activity_search(request)
     return render(request, "users/activity.html")
-
-
-# @login_required
-# def users_list(request):
-#     users = Profile.objects.exclude(user=request.user)
-#     sent_friend_requests = FriendRequest.objects.filter(from_user=request.user)
-#     sent_to = []
-#     friends = []
-#     for user in users:
-#         friend = user.friends.all()
-#         for f in friend:
-#             if f in friends:
-#                 friend = friend.exclude(user=f.user)
-#         friends += friend
-#     my_friends = request.user.profile.friends.all()
-#     for i in my_friends:
-#         if i in friends:
-#             friends.remove(i)
-#     if request.user.profile in friends:
-#         friends.remove(request.user.profile)
-#     random_list = random.sample(list(users), min(len(list(users)), 10))
-#     for r in random_list:
-#         if r in friends:
-#             random_list.remove(r)
-#     friends += random_list
-#     for i in my_friends:
-#         if i in friends:
-#             friends.remove(i)
-#     for se in sent_friend_requests:
-#         sent_to.append(se.to_user)
-#     context = {'users': friends, 'sent': sent_to}
-#     return render(request, 'users/users_list.html', context)
-#
-#
-# def friend_list(request):
-#     p = request.user.profile
-#     friends = p.friends.all()
-#     context = {'friends': friends}
-#     return render(request, 'users/friend_list.html', context)
-#
-#
-# @login_required
-# def send_friend_request(request, id):
-#     user = get_object_or_404(User, id=id)
-#     frequest, created = FriendRequest.objects.get_or_create(
-#         from_user=request.user, to_user=user
-#     )
-#     return HttpResponseRedirect('/users/{}'.format(user.profile.slug))
-#
-#
-# @login_required
-# def cancel_friend_request(request, id):
-#     user = get_object_or_404(User, id=id)
-#     frequest = FriendRequest.objects.filter(
-#         from_user=request.user, to_user=user
-#     ).first()
-#     frequest.delete()
-#     return HttpResponseRedirect('/users/{}'.format(user.profile.slug))
-#
-#
-# @login_required
-# def accept_friend_request(request, id):
-#     from_user = get_object_or_404(User, id=id)
-#     frequest = FriendRequest.objects.filter(
-#         from_user=from_user, to_user=request.user
-#     ).first()
-#     user1 = frequest.to_user
-#     user2 = from_user
-#     user1.profile.friends.add(user2.profile)
-#     user2.profile.friends.add(user1.profile)
-#     if FriendRequest.objects.filter(from_user=request.user, to_user=from_user).first():
-#         request_rev = FriendRequest.objects.filter(
-#             from_user=request.user, to_user=from_user
-#         ).first()
-#         request_rev.delete()
-#     frequest.delete()
-#     return HttpResponseRedirect('/users/{}'.format(request.user.profile.slug))
-#
-#
-# @login_required
-# def delete_friend_request(request, id):
-#     from_user = get_object_or_404(User, id=id)
-#     frequest = FriendRequest.objects.filter(
-#         from_user=from_user, to_user=request.user
-#     ).first()
-#     frequest.delete()
-#     return HttpResponseRedirect('/users/{}'.format(request.user.profile.slug))
-#
-#
-# def delete_friend(request, id):
-#     user_profile = request.user.profile
-#     friend_profile = get_object_or_404(Profile, id=id)
-#     user_profile.friends.remove(friend_profile)
-#     friend_profile.friends.remove(user_profile)
-#     return HttpResponseRedirect('/users/{}'.format(friend_profile.slug))
-#
-#
-# @login_required
-# def profile_view(request, slug):
-#     p = Profile.objects.filter(slug=slug).first()
-#     u = p.user
-#     sent_friend_requests = FriendRequest.objects.filter(from_user=p.user)
-#     rec_friend_requests = FriendRequest.objects.filter(to_user=p.user)
-#
-#     friends = p.friends.all()
-#
-#     # is this user our friend
-#     button_status = 'none'
-#     if p not in request.user.profile.friends.all():
-#         button_status = 'not_friend'
-#
-#         # if we have sent him a friend request
-#         if (
-#             len(
-#                 FriendRequest.objects.filter(from_user=request.user).filter(
-#                     to_user=p.user
-#                 )
-#             )
-#             == 1
-#         ):
-#             button_status = 'friend_request_sent'
-#
-#         # if we have recieved a friend request
-#         if (
-#             len(
-#                 FriendRequest.objects.filter(from_user=p.user).filter(
-#                     to_user=request.user
-#                 )
-#             )
-#             == 1
-#         ):
-#             button_status = 'friend_request_received'
-#
-#     context = {
-#         'u': u,
-#         'button_status': button_status,
-#         'friends_list': friends,
-#         'sent_friend_requests': sent_friend_requests,
-#         'rec_friend_requests': rec_friend_requests,
-#     }
-#
-#     return render(request, 'users/edit_profile.html', context)
-#
-#
-# @login_required
-# def edit_profile(request):
-#     if request.method == 'POST':
-#         u_form = UserUpdateForm(request.POST, instance=request.user)
-#         p_form = ProfileUpdateForm(
-#             request.POST, request.FILES, instance=request.user.profile
-#         )
-#         if u_form.is_valid() and p_form.is_valid():
-#             u_form.save()
-#             p_form.save()
-#             messages.success(request, f'Your account has been updated!')
-#             return redirect('my_profile')
-#     else:
-#         u_form = UserUpdateForm(instance=request.user)
-#         p_form = ProfileUpdateForm(instance=request.user.profile)
-#     context = {
-#         'u_form': u_form,
-#         'p_form': p_form,
-#     }
-#     return render(request, 'users/edit_profile.html', context)
-#
-#
-# @login_required
-# def my_profile(request):
-#     p = request.user.profile
-#     you = p.user
-#     sent_friend_requests = FriendRequest.objects.filter(from_user=you)
-#     rec_friend_requests = FriendRequest.objects.filter(to_user=you)
-#     friends = p.friends.all()
-#
-#     # is this user our friend
-#     button_status = 'none'
-#     if p not in request.user.profile.friends.all():
-#         button_status = 'not_friend'
-#
-#         # if we have sent him a friend request
-#         if (
-#             len(
-#                 FriendRequest.objects.filter(from_user=request.user).filter(to_user=you)
-#             )
-#             == 1
-#         ):
-#             button_status = 'friend_request_sent'
-#
-#         if (
-#             len(
-#                 FriendRequest.objects.filter(from_user=p.user).filter(
-#                     to_user=request.user
-#                 )
-#             )
-#             == 1
-#         ):
-#             button_status = 'friend_request_received'
-#
-#     context = {
-#         'u': you,
-#         'button_status': button_status,
-#         'friends_list': friends,
-#         'sent_friend_requests': sent_friend_requests,
-#         'rec_friend_requests': rec_friend_requests,
-#     }
-#
-#     return render(request, 'users/edit_profile.html', context)
-#
-#
-# @login_required
-# def search_users(request):
-#     query = request.GET.get('q')
-#     object_list = User.objects.filter(username__icontains=query)
-#     context = {'users': object_list}
-#     return render(request, 'users/search_users.html', context)
